+++ zpy_addon/__init__.py
@@ -13,6 +13,7 @@ try:
 except ModuleNotFoundError:
     log.warn('Could not find required pip packages.')
     install_pip_depenencies()
+    import zpy
 
 bl_info = {
     "name": "zpy",
@@ -29,42 +30,45 @@ bl_info = {
 
 if "bpy" in locals():
     log.warning('Reloading zpy_addon files.')
-    from . import export, render, script, segment
-    importlib.reload(export)
-    importlib.reload(render)
-    importlib.reload(script)
-    importlib.reload(segment)
+    from . import export_panel
+    from . import render_panel
+    from . import script_panel
+    from . import segment_panel
+    importlib.reload(export_panel)
+    importlib.reload(render_panel)
+    importlib.reload(script_panel)
+    importlib.reload(segment_panel)
     importlib.reload(zpy)
 
 classes = (
     # Properties
-    segment.CategoryProperties,
-    segment.SegmentableProperties,
+    segment_panel.CategoryProperties,
+    segment_panel.SegmentableProperties,
     # Object operators
-    segment.SegmentInstanceSingle,
-    segment.SegmentInstanceMany,
-    segment.ResetSegData,
+    segment_panel.SegmentInstanceSingle,
+    segment_panel.SegmentInstanceMany,
+    segment_panel.ResetSegData,
     # Scene operators
-    segment.VisualizeInstance,
-    segment.VisualizeCategory,
-    segment.CategoriesFromText,
-    segment.CategoriesFromZUMOJSON,
-    render.RenderOperator,
-    render.OpenOutputDirOperator,
-    render.CleanOutputDirOperator,
-    export.ExportOperator,
-    export.OpenExportDirOperator,
-    export.CleanUpDirOperator,
-    script.LoadGinConfigOperator,
-    script.PushGinConfigOperator,
-    script.LoadRunpyOperator,
-    script.PushRunpyOperator,
-    script.RunOperator,
+    segment_panel.VisualizeInstance,
+    segment_panel.VisualizeCategory,
+    segment_panel.CategoriesFromText,
+    segment_panel.CategoriesFromZUMOJSON,
+    render_panel.RenderOperator,
+    render_panel.OpenOutputDirOperator,
+    render_panel.CleanOutputDirOperator,
+    export_panel.ExportOperator,
+    export_panel.OpenExportDirOperator,
+    export_panel.CleanUpDirOperator,
+    script_panel.LoadGinConfigOperator,
+    script_panel.PushGinConfigOperator,
+    script_panel.LoadRunpyOperator,
+    script_panel.PushRunpyOperator,
+    script_panel.RunOperator,
     # Panels
-    segment.SegmentPanel,
-    render.RenderPanel,
-    script.ScriptPanel,
-    export.ExportPanel,
+    segment_panel.SegmentPanel,
+    render_panel.RenderPanel,
+    script_panel.ScriptPanel,
+    export_panel.ExportPanel,
 )
 
 
@@ -76,11 +80,11 @@ def register():
             bpy.utils.register_class(cls)
         except Exception as e:
             log.warning(f'Exception when registering {cls.__name__}: {e}')
-    segment.registerObjectProperties()
-    segment.registerSceneProperties()
-    render.registerSceneProperties()
-    export.registerSceneProperties()
-    script.registerSceneProperties()
+    segment_panel.registerObjectProperties()
+    segment_panel.registerSceneProperties()
+    render_panel.registerSceneProperties()
+    export_panel.registerSceneProperties()
+    script_panel.registerSceneProperties()
 
 
 def unregister():
@@ -95,6 +99,8 @@ def unregister():
 
 def install_pip_depenencies():
     """ Install pip dependencies required by zpy addon."""
+    log.info('Installing zpy and it\'s pip dendencies.')
+    
     import subprocess
     import sys
     # TODO: METHOD 1
