+++ zpy_addon/segment_panel.py
